DBIResultArrow-class {DBI}	R Documentation

DBIResultArrow class

Description

This virtual class describes the result and state of execution of a DBMS statement (any statement, query or non-query) for returning data as an Arrow object.

Implementation notes

Individual drivers are free to allow single or multiple active results per connection.

The default show method displays a summary of the query using other DBI generics.

See Also

Other DBI classes: DBIConnection-class, DBIConnector-class, DBIDriver-class, DBIObject-class, DBIResult-class

Other DBIResultArrow generics: dbBind(), dbClearResult(), dbFetchArrow(), dbFetchArrowChunk(), dbHasCompleted(), dbIsValid()

[Package DBI version 1.2.2 Index]