DBI specification

The DBI package defines the generic DataBase Interface for R. The connection to individual DBMS is made by packages that import DBI (so-called DBI backends). This document formalizes the behavior expected by the functions declared in DBI and implemented by the individal backends.

To ensure maximum portability and exchangeability, and to reduce the effort for implementing a new DBI backend, the DBItest package defines a comprehensive set of test cases that test conformance to the DBI specification. In fact, this document is derived from comments in the test definitions of the DBItest package. This ensures that an extension or update to the tests will be reflected in this document.

Getting started

A DBI backend is an R package, which should import the DBI and methods packages. For better or worse, the names of many existing backends start with ‘R’, e.g., RSQLite, RMySQL, RSQLServer; it is up to the package author to adopt this convention or not.

DBI classes and methods

A backend defines three classes, which are subclasses of DBIDriver, DBIConnection, and DBIResult. The backend provides implementation for all methods defined for these base classes, but not provided by DBI. All methods have an ellipsis ... in their formals.

Construction of the <code>DBIDriver</code> object

The backend must support creation of an instance of its DBIDriver subclass with a constructor function. By default, its name is the package name without the leading ‘R’ (if it exists), e.g., SQLite for the RSQLite package. For the automated tests, the constructor name can be tweaked using the constructor_name tweak.

The constructor must be exported, and it must be a function that is callable without arguments. For the automated tests, unless the constructor_relax_args tweak is set to TRUE, an empty argument list is expected. Otherwise, an argument list where all arguments have default values is also accepted.

Specification

The backend can override the dbDataType() generic for its driver class.

This generic expects an arbitrary object as second argument. To query the values returned by the default implementation, run example(dbDataType, package = "DBI"). If the backend needs to override this generic, it must accept all basic R data types as its second argument, namely logical, integer, numeric, character, dates (see Dates), date-time (see DateTimeClasses), and difftime. It also must accept lists of raw vectors and map them to the BLOB (binary large object) data type, unless the omit_blob_tests tweak is set to TRUE. As-is objects (i.e., wrapped by I()) must be supported and return the same results as their unwrapped counterparts. The SQL data type for factor and ordered is the same as for character. The behavior for other object types is not specified.

All data types returned by dbDataType() are usable in an SQL statement of the form "CREATE TABLE test (a ...)".

DBI specifies only the return type for dbConnect().

A warning is issued when releasing a connection without calling dbDisconnect(), or when calling dbDisconnect() on an already disconnected connection.

No warnings occur under normal conditions. The DBIResult object returned by dbSendQuery() must be valid, i.e., dbIsValid() returns TRUE. When done, the DBIResult object must be cleared with a call to dbClearResult(). Failure to clear the result set leads to a warning when the connection is closed.

If the backend supports only one open result set per connection, issuing a second query invalidates an already open result set and raises a warning. The newly opened result set is valid and must be cleared with dbClearResult().

Fetching multi-row queries with one or more columns be default returns the entire result. Multi-row queries can also be fetched progressively by passing a whole number (integer or numeric) as the n argument. A value of Inf for the n argument is supported and also returns the full result. If more rows than available are fetched, the result is returned in full without warning. If zero rows are fetched, the columns of the data frame are still fully typed. Fetching fewer rows than available is permitted, no warning is issued when clearing the result set.

The column types of the returned data frame depend on the data returned:

integer for integer values between -2^31 and 2^31 - 1

numeric for numbers with a fractional component

logical for Boolean values (some backends may return an integer)

character for text

lists of raw for blobs (with NULL entries for SQL NULL values)

coercible using as.Date() for dates (also applies to the return value of the SQL function current_date)

coercible using hms::as.hms() for times (also applies to the return value of the SQL function current_time)

coercible using as.POSIXct() for timestamps (also applies to the return value of the SQL function current_timestamp)

NA for SQL NULL values

If dates and timestamps are supported by the backend, the following R types are used:

Date for dates (also applies to the return value of the SQL function current_date)

POSIXct for timestamps (also applies to the return value of the SQL function current_timestamp)

R has no built-in type with lossless support for the full range of 64-bit or larger integers. Here, the following rules apply:

Values are returned as numeric

Conversion to character always returns a lossless decimal representation of the data

dbClearResult() frees all resources associated with retrieving the result of a query or update operation. The DBI backend can expect a call to dbClearResult() for each dbSendQuery() or dbSendStatement() call.

Fetching multi-row queries with one or more columns be default returns the entire result. A value of Inf for the n argument is supported and also returns the full result. If more rows than available are fetched, the result is returned in full without warning. If zero rows are fetched, the columns of the data frame are still fully typed. Fetching fewer rows than available is permitted, no warning is issued.

No warnings occur under normal conditions. The DBIResult object returned by dbSendStatement() must be valid, i.e., dbIsValid() returns TRUE. When done, the DBIResult object must be cleared with a call to dbClearResult(). Failure to clear the result set leads to a warning when the connection is closed. If the backend supports only one open result set per connection, issuing a second query invalidates an already open result set and raises a warning. The newly opened result set is valid and must be cleared with dbClearResult().

No warnings occur under normal conditions.

The returned expression can be used in a SELECT ... query, and for any scalar character x the value of dbGetQuery(paste0("SELECT ", dbQuoteString(x)))[[1]] must be identical to x, even if x contains spaces, quotes (single or double), backticks, or newlines (in any combination) or is itself the result of a dbQuoteString() call coerced back to character (even repeatedly). If x is NA, the result must merely satisfy is.na(). The strings "NA" or "NULL" are not treated specially.

NA should be translated to an unquoted SQL NULL, so that the query SELECT * FROM (SELECT 1) a WHERE ... IS NULL returns one row.

Calling dbGetQuery() for a query of the format SELECT 1 AS ... returns a data frame with the identifier, unquoted, as column name. Quoted identifiers can be used as table and column names in SQL queries, in particular in queries like SELECT 1 AS ... and SELECT * FROM (SELECT 1) .... This is also true for column names that are empty strings or contain special characters such as a space, a dot, a comma, or quotes used to mark strings or identifiers, if the database supports this. In any case, checking the validity of the identifier should be performed only when executing a query, and not by dbQuoteIdentifier().

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbReadTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name, ...) so that all optional arguments are passed along

If the result of a call to dbQuoteIdentifier(): no more quoting is done

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbWriteTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name, ...) so that all optional arguments are passed along

If the result of a call to dbQuoteIdentifier(): no more quoting is done

If the overwrite argument is TRUE, an existing table of the same name will be overwritten. This argument doesn't change behavior if the table does not exist yet.

If the append argument is TRUE, the rows in an existing table are preserved, and the new data are appended. If the table doesn't exist yet, it is created.

If the temporary argument is TRUE, the table is not available in a second connection and is gone after reconnecting. Not all backends support this argument. A regular, non-temporary table is visible in a second connection and after reconnecting to the database.

SQL keywords can be used freely in table names, column names, and data. Quotes, commas, and spaces can also be used in the data, and, if the database supports non-syntactic identifiers, also for table names and column names.

The following data types must be supported at least, and be read identically with dbReadTable():

integer

numeric (also with Inf and NaN values, the latter are translated to NA)

logical

NA as NULL

64-bit values (using "bigint" as field type)

character (in both UTF-8 and native encodings)

factor (returned as character)

list of raw (if supported by the database)

date (if supported by the database; returned as integers with class Date)

time (if supported by the database; returned as objects that inherit from difftime)

timestamp (if supported by the database; with time zone support)

The interpretation of rownames depends on the row.names argument, see sqlRownamesToColumn() for details:

If FALSE, row names are ignored.

If TRUE, row names are converted to a column named "row_names", even if the input data frame only has natural row names from 1 to nrow(...).

If NA, a column named "row_names" is created if the data has custom row names, no extra column is created in the case of natural row names.

If a string, this specifies the name of the column in the remote table that contains the row names, even if the input data frame only has natural row names.

TBD (schema support).

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbExistsTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name, ...) so that all optional arguments are passed along

If the result of a call to dbQuoteIdentifier(): no more quoting is done

For all tables listed by dbListTables(), dbExistsTable() returns TRUE.

A table removed by dbRemoveTable() doesn't appear in the list of tables returned by dbListTables(), and dbExistsTable() returns FALSE. The removal propagates immediately to other connections to the same database. This function can also be used to remove a temporary table.

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbRemoveTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name, ...) so that all optional arguments are passed along

If the result of a call to dbQuoteIdentifier(): no more quoting is done

The completion status for a query is only guaranteed to be set to FALSE after attempting to fetch past the end of the entire result. Therefore, for a query with an empty result set, the initial return value is unspecified, but the result value is TRUE after trying to fetch only one row. Similarly, for a query with a result set of length n, the return value is unspecified after fetching n rows, but the result value is TRUE after trying to fetch only one more row.

DBI clients execute parametrized statements as follows:

Call dbSendQuery() or dbSendStatement() with a query or statement that contains placeholders, store the returned DBIResult object in a variable. Mixing placeholders (in particular, named and unnamed ones) is not recommended. It is good practice to register a call to dbClearResult() via on.exit() right after calling dbSendQuery() or dbSendStatement() (see the last enumeration item).

Construct a list with parameters that specify actual values for the placeholders. The list must be named or unnamed, depending on the kind of placeholders used. Named values are matched to named parameters, unnamed values are matched by position in the list of parameters. All elements in this list must have the same lengths and contain values supported by the backend; a data.frame is internally stored as such a list. The parameter list is passed to a call to dbBind() on the DBIResult object.

Retrieve the data or the number of affected rows from the DBIResult object.

For queries issued by dbSendQuery(), call dbFetch().

For statements issued by dbSendStatements(), call dbGetRowsAffected(). (Execution begins immediately after the dbBind() call, the statement is processed entirely before the function returns.)

Repeat 2. and 3. as necessary.

Close the result set via dbClearResult().

The elements of the params argument do not need to be scalars, vectors of arbitrary length (including length 0) are supported. For queries, calling dbFetch() binding such parameters returns concatenated results, equivalent to binding and fetching for each set of values and connecting via rbind(). For data manipulation statements, dbGetRowsAffected() returns the total number of rows affected if binding non-scalar parameters. dbBind() also accepts repeated calls on the same result set for both queries and data manipulation statements, even if no results are fetched between calls to dbBind().

At least the following data types are accepted:

integer

numeric

logical for Boolean values (some backends may return an integer)

NA

character

Date

POSIXct timestamps

POSIXlt timestamps

lists of raw for blobs (with NULL entries for SQL NULL values)

Actual support for transactions may vary between backends. A transaction is initiated by a call to dbBegin() and committed by a call to dbCommit(). Data written in a transaction must persist after the transaction is committed. For example, a table that is missing when the transaction is started but is created and populated during the transaction must exist and contain the data added there both during and after the transaction, and also in a new connection.

A transaction can also be aborted with dbRollback(). All data written in such a transaction must be removed after the transaction is rolled back. For example, a table that is missing when the transaction is started but is created during the transaction must not exist anymore after the rollback.

Disconnection from a connection with an open transaction gives a warning and effectively rolls back the transaction. All data written in such a transaction must be removed after the transaction is rolled back.

The behavior is not specified if other arguments are passed to these functions. In particular, RSQLite issues named transactions if the name argument is set.

The transaction isolation level is not specified by DBI.

dbWithTransaction() initiates a transaction with dbBegin(), executes the code given in the code argument, and commits the transaction with dbCommit(). If the code raises an error, the transaction is instead aborted with dbRollback(), and the error is propagated. If the code calls dbBreak(), execution of the code stops and the transaction is silently aborted. All side effects caused by the code (such as the creation of new variables) propagate to the calling environment.

Additional arguments

The following arguments are not part of the dbReadTable() generic (to improve compatibility across backends) but are part of the DBI specification:

row.names

check.names

They must be provided as named arguments. See the "Value" section for details on their usage.

The following arguments are not part of the dbWriteTable() generic (to improve compatibility across backends) but are part of the DBI specification:

row.names (default: NA)

overwrite (default: FALSE)

append (default: FALSE)

field.types (default: NULL)

temporary (default: FALSE)

They must be provided as named arguments. See the "Specification" and "Value" sections for details on their usage.

TBD (schema support).

TBD: Schema support.

They must be provided as named arguments. See the "Specification" section for details on their usage.

TBD: Schema support.

They must be provided as named arguments. See the "Specification" section for details on their usage.