Table-class {DBI}	R Documentation

Refer to a table nested in a hierarchy (e.g. within a schema)

Description

Refer to a table nested in a hierarchy (e.g. within a schema)

Usage

Table(...)

Arguments

...
Components of the hierarchy, e.g. schema, table, or cluster, catalog, schema, table. For more on these concepts, see http://stackoverflow.com/questions/7022755/

[Package DBI version 0.5 Index]