Disconnect (close) a connection

This closes the connection, discards all pending work, and frees resources (e.g., memory, sockets).

Methods in other packages

RSQLite::dbDisconnect("SQLiteConnection")

dbDisconnect(conn, ...)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

...
Other parameters passed on to methods.

Value

dbDisconnect() returns TRUE, invisibly.

Specification

A warning is issued on garbage collection when a connection has been released without calling dbDisconnect(), but this cannot be tested automatically. A warning is issued immediately when calling dbDisconnect() on an already disconnected or invalid connection.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbDisconnect(con)