dbGetConnectArgs {DBI}	R Documentation

Get connection arguments

Description

Returns the arguments stored in a DBIConnector object for inspection, optionally evaluating them. This function is called by dbConnect() and usually does not need to be called directly.

Usage

dbGetConnectArgs(drv, eval = TRUE, ...)

Arguments

drv
A object inheriting from DBIConnector.

eval
Set to FALSE to return the functions that generate the argument instead of evaluating them.

...
Other arguments passed on to methods. Not otherwise used.

See Also

Other DBIConnector generics: DBIConnector-class, dbConnect(), dbDataType(), dbIsReadOnly()

Examples

cnr <- new("DBIConnector",
  .drv = RSQLite::SQLite(),
  .conn_args = list(dbname = ":memory:", password = function() "supersecret")
)
dbGetConnectArgs(cnr)
dbGetConnectArgs(cnr, eval = FALSE)

[Package DBI version 1.1.3 Index]