Quote literal values

Call these methods to generate a string that is suitable for use in a query as a literal value of the correct type, to make sure that you generate valid SQL and protect against SQL injection attacks.

dbQuoteLiteral(conn, x, ...)

Arguments

conn
A subclass of DBIConnection, representing an active connection to an DBMS.

x
A vector to quote as string.

...
Other arguments passed on to methods.

See also

Examples

# Quoting ensures that arbitrary input is safe for use in a query
name <- "Robert'); DROP TABLE Students;--"
dbQuoteLiteral(ANSI(), name)

#> <SQL> 'Robert''); DROP TABLE Students;--'

# NAs become NULL
dbQuoteLiteral(ANSI(), c(1:3, NA))

#> <SQL> 1
#> <SQL> 2
#> <SQL> 3
#> <SQL> NULL

# Logicals become integers by default
dbQuoteLiteral(ANSI(), c(TRUE, FALSE, NA))

#> <SQL> 1
#> <SQL> 0
#> <SQL> NULL

# Raw vectors become hex strings by default
dbQuoteLiteral(ANSI(), list(as.raw(1:3), NULL))

#> <SQL> X'010203'
#> <SQL> NULL

# SQL vectors are always passed through as is
var_name <- SQL("select")
var_name

#> <SQL> select

dbQuoteLiteral(ANSI(), var_name)

#> <SQL> select

# This mechanism is used to prevent double escaping
dbQuoteLiteral(ANSI(), dbQuoteLiteral(ANSI(), name))

#> <SQL> 'Robert''); DROP TABLE Students;--'