Remove a table from the database

Remove a remote table (e.g., created by dbWriteTable()) from the database.

dbRemoveTable(conn, name, ...)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

name

A character string specifying a DBMS table name.

...

Other parameters passed on to methods.

Value

dbRemoveTable() returns TRUE, invisibly. If the table does not exist, an error is raised. An attempt to remove a view with this function may result in an error.

An error is raised if name cannot be processed with dbQuoteIdentifier() or if this results in a non-scalar.

Additional arguments

TBD: Schema support.

They must be provided as named arguments. See the "Specification" section for details on their usage.

Specification

A table removed by dbRemoveTable() doesn't appear in the list of tables returned by dbListTables(), and dbExistsTable() returns FALSE. The removal propagates immediately to other connections to the same database. This function can also be used to remove a temporary table.

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbRemoveTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name, ...) so that all optional arguments are passed along

If the result of a call to dbQuoteIdentifier(): no more quoting is done

See also

Other DBIConnection generics: DBIConnection-class, dbDataType, dbDisconnect, dbExecute, dbExistsTable, dbGetException, dbGetInfo, dbGetQuery, dbIsValid, dbListFields, dbListResults, dbListTables, dbReadTable, dbSendQuery, dbSendStatement, dbWriteTable

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbExistsTable(con, "iris")

#> [1] FALSE

dbWriteTable(con, "iris", iris)

#> [1] TRUE

dbExistsTable(con, "iris")

#> [1] TRUE

dbRemoveTable(con, "iris")
dbExistsTable(con, "iris")

#> [1] FALSE

dbDisconnect(con)