Unquote identifiers

Call this method to convert a SQL object created by dbQuoteIdentifier() back to a list of Table objects.
Methods in other packages

DBI::dbUnquoteIdentifier("DBIConnection")

dbUnquoteIdentifier(conn, x, ...)

Arguments

conn
A subclass of DBIConnection, representing an active connection to an DBMS.

x
An SQL or Table object or character vector, or a list of such objects, to unquote.

...
Other arguments passed on to methods.

See also

Other DBIResult generics: DBIResult-class, dbBind, dbClearResult, dbColumnInfo, dbFetch, dbGetInfo, dbGetRowCount, dbGetRowsAffected, dbGetStatement, dbHasCompleted, dbIsValid, dbQuoteIdentifier, dbQuoteLiteral, dbQuoteString

Examples

# Unquoting allows to understand the structure of a possibly complex quoted
# identifier

dbUnquoteIdentifier(
  ANSI(),
  SQL(c('"Schema"."Table"', "UnqualifiedTable"))
)

#> Error in match.fun(FUN): object 'Table' not found

# Character vectors are wrapped in a list
dbQuoteIdentifier(
  ANSI(),
  c(schema = "Schema", table = "Table")
)

#> <SQL> "Schema"
#> <SQL> "Table"

# Lists of character vectors are returned unchanged
dbQuoteIdentifier(
  ANSI(),
  list(c(schema = "Schema", table = "Table"), "UnqualifiedTable")
)

#> Error: x must be character or SQL