Self-contained SQL transactions

Given that transactions are implemented, this function allows you to pass in code that is run in a transaction. The default method of dbWithTransaction calls dbBegin() before executing the code, and dbCommit() after successful completion, or dbRollback() in case of an error. The advantage is that you don't have to remember to do dbBegin and dbCommit or dbRollback -- that is all taken care of. The special function dbBreak allows an early exit with rollback, it can be called only inside dbWithTransaction.

dbWithTransaction(conn, code, ...)

dbBreak()

Arguments

conn

A DBIConnection object, as returned by dbConnect().

code

An arbitrary block of R code

...

Other parameters passed on to methods.

Value

The result of the evaluation of code

Side Effects

The transaction in code on the connection conn is committed or rolled back. The code chunk may also modify the local R environment.

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "cars", head(cars, 3))

#> [1] TRUE

dbReadTable(con, "cars")   # there are 3 rows

#>   speed dist
#> 1     4    2
#> 2     4   10
#> 3     7    4

## successful transaction
dbWithTransaction(con, {
  dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (1, 1);")
  dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (2, 2);")
  dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (3, 3);")
})

#> [1] 1

dbReadTable(con, "cars")   # there are now 6 rows

#>   speed dist
#> 1     4    2
#> 2     4   10
#> 3     7    4
#> 4     1    1
#> 5     2    2
#> 6     3    3

## failed transaction -- note the missing comma
tryCatch(
  dbWithTransaction(con, {
    dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (1, 1);")
    dbExecute(con, "INSERT INTO cars (speed dist) VALUES (2, 2);")
    dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (3, 3);")
  }),
  error = identity
)

#> <Rcpp::exception in rsqlite_send_query(conn@ptr, statement): near "dist": syntax error>

dbReadTable(con, "cars")   # still 6 rows

#>   speed dist
#> 1     4    2
#> 2     4   10
#> 3     7    4
#> 4     1    1
#> 5     2    2
#> 6     3    3

## early exit, silently
dbWithTransaction(con, {
  dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (1, 1);")
  dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (2, 2);")
  if (nrow(dbReadTable(con, "cars")) > 7) dbBreak()
  dbExecute(con, "INSERT INTO cars (speed, dist) VALUES (3, 3);")
})
dbReadTable(con, "cars")   # still 6 rows

#>   speed dist
#> 1     4    2
#> 2     4   10
#> 3     7    4
#> 4     1    1
#> 5     2    2
#> 6     3    3

dbDisconnect(con)