DBI: R Database Interface

A database interface definition for communication between R and relational database management systems. All classes in this package are virtual and need to be extended by the various R/DBMS implementations.

See also

Important generics: dbConnect(), dbGetQuery(), dbWriteTable(), dbDisconnect()

Formal specification (currently work in progress and incomplete): DBIspec

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "iris", iris)

#> [1] TRUE

dbGetQuery(con, "SELECT * FROM iris WHERE [Petal.Width] > 2.3")

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 1          6.3         3.3          6.0         2.5 virginica
#> 2          7.2         3.6          6.1         2.5 virginica
#> 3          5.8         2.8          5.1         2.4 virginica
#> 4          6.3         3.4          5.6         2.4 virginica
#> 5          6.7         3.1          5.6         2.4 virginica
#> 6          6.7         3.3          5.7         2.5 virginica

dbDisconnect(con)