DBI-package {DBI}	R Documentation

R Database Interface

Description

A database interface definition for communication between R and relational database management systems. All classes in this package are virtual and need to be extended by the various R/DBMS implementations.

See Also

Important generics: dbConnect, dbGetQuery, dbWriteTable, dbDisconnect

Formal specification (currently work in progress and incomplete): DBIspec

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "iris", iris)
dbGetQuery(con, "SELECT * FROM iris WHERE [Petal.Width] > 2.3")

dbDisconnect(con)

[Package DBI version 0.5-1 Index]