DBIConnection-class {DBI}	R Documentation

DBIConnection class.

Description

This virtual class encapsulates the connection to a DBMS, and it provides access to dynamic queries, result sets, DBMS session management (transactions), etc.

Usage

## S4 method for signature 'DBIConnection'
show(object)

Arguments

object
Object to display

Implementation note

Individual drivers are free to implement single or multiple simultaneous connections.

See Also

Other DBI classes: DBIDriver-class, DBIObject-class, DBIResult-class

Examples

## Not run: 
con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbDisconnect(con)

con <- dbConnect(RPostgreSQL::PostgreSQL(), "username", "passsword")
dbDisconnect(con)

## End(Not run)

[Package DBI version 0.4-1 Index]