DBIResult-class {DBI}	R Documentation

DBIResult class.

Description

This virtual class describes the result and state of execution of a DBMS statement (any statement, query or non-query). The result set keeps track of whether the statement produces output how many rows were affected by the operation, how many rows have been fetched (if statement is a query), whether there are more rows to fetch, etc.

Usage

## S4 method for signature 'DBIResult'
show(object)

Arguments

object
Object to display

Implementation notes

Individual drivers are free to allow single or multiple active results per connection.

The default show method displays a summary of the query using other DBI generics.

See Also

Other DBI classes: DBIConnection-class, DBIDriver-class, DBIObject-class

[Package DBI version 0.4-1 Index]