Id-class {DBI}	R Documentation

Refer to a table nested in a hierarchy (e.g. within a schema)

Description

Objects of class Id have a single slot name, which is a named character vector.

Usage

Id(...)

Arguments

...
Components of the hierarchy, e.g. schema, table, or cluster, catalog, schema, table. For more on these concepts, see http://stackoverflow.com/questions/7022755/

[Package DBI version 1.0.0 Index]