dbAppendTable {DBI}	R Documentation

Insert rows into a table

Description

The dbAppendTable() method assumes that the table has been created beforehand, e.g. with dbCreateTable(). The default implementation calls sqlAppendTableTemplate() and then dbExecute() with the param argument. Backends compliant to ANSI SQL 99 which use ? as a placeholder for prepard queries don't need to override it. Backends with a different SQL syntax which use ? as a placeholder for prepared queries can override sqlAppendTable(). Other backends (with different placeholders or with entirely different ways to create tables) need to override the dbAppendTable() method.

Usage

dbAppendTable(conn, name, value, ..., row.names = NULL)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

name
Name of the table, escaped with dbQuoteIdentifier().

value
A data frame of values. The column names must be consistent with those in the target table in the database.

...
Other arguments used by individual methods.

row.names
Must be NULL.

Details

The row.names argument is not supported by this method. Process the values with sqlRownamesToColumn() before calling this method.

See Also

Other DBIConnection generics: DBIConnection-class, dbCreateTable, dbDataType, dbDisconnect, dbExecute, dbExistsTable, dbGetException, dbGetInfo, dbGetQuery, dbIsReadOnly, dbIsValid, dbListFields, dbListObjects, dbListResults, dbListTables, dbReadTable, dbRemoveTable, dbSendQuery, dbSendStatement, dbWriteTable

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbCreateTable(con, "iris", iris)
dbAppendTable(con, "iris", iris)
dbReadTable(con, "iris")
dbDisconnect(con)

[Package DBI version 1.0.0 Index]