dbCallProc {DBI}	R Documentation

Call an SQL stored procedure

Description

Deprecated since 2014

Usage

dbCallProc(conn, ...)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

...
Other parameters passed on to methods.

Details

The recommended way of calling a stored procedure is now

dbGetQuery if a result set is returned

dbExecute for data manipulation and other cases where no result set is returned

[Package DBI version 1.1.2 Index]