dbClearResult {DBI}	R Documentation

Clear a result set.

Description

Frees all resources (local and remote) associated with a result set. It some cases (e.g., very large result sets) this can be a critical step to avoid exhausting resources (memory, file descriptors, etc.)

Usage

dbClearResult(res, ...)

Arguments

res
An object inheriting from DBIResult.

...
Other arguments passed on to methods.

Value

a logical indicating whether clearing the result set was successful or not.

See Also

Other DBIResult generics: dbColumnInfo, dbFetch, dbGetRowCount, dbGetRowsAffected, dbGetStatement, dbHasCompleted

[Package DBI version 0.4-1 Index]