Clear a result set

Frees all resources (local and remote) associated with a result set. In some cases (e.g., very large result sets) this can be a critical step to avoid exhausting resources (memory, file descriptors, etc.)

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

RAthena::dbClearResult("AthenaResult")

bigrquery::dbClearResult("BigQueryResult")

RClickhouse::dbClearResult("ClickhouseResult")

DatabaseConnector::dbClearResult("DatabaseConnectorResult")

dittodb::dbClearResult("DBIMockResult")

sparklyr::dbClearResult("DBISparkResult")

sergeant::dbClearResult("DrillResult")

duckdb::dbClearResult("duckdb_result")

RJDBC::dbClearResult("JDBCResult")

RMariaDB::dbClearResult("MariaDBResult")

RMySQL::dbClearResult("MySQLResult")

odbc::dbClearResult("OdbcResult")

RPostgreSQL::dbClearResult("PostgreSQLResult")

RPostgres::dbClearResult("PqResult")

RPresto::dbClearResult("PrestoResult")

lazysf::dbClearResult("SFSQLResult")

RSQLite::dbClearResult("SQLiteResult")

dbClearResult(res, ...)

Arguments

res

An object inheriting from DBIResult.

...

Other arguments passed on to methods.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

rs <- dbSendQuery(con, "SELECT 1")
print(dbFetch(rs))
#>   1
#> 1 1

dbClearResult(rs)
dbDisconnect(con)