Clear a result set

Frees all resources (local and remote) associated with a result set. This step is mandatory for all objects obtained by calling dbSendQuery() or dbSendStatement().

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

bigrquery::dbClearResult("BigQueryResult")

DatabaseConnector::dbClearResult("DatabaseConnectorDbiResult")

DatabaseConnector::dbClearResult("DatabaseConnectorJdbcResult")

dittodb::dbClearResult("DBIMockResult")

duckdb::dbClearResult("duckdb_result")

lazysf::dbClearResult("SFSQLResult")

odbc::dbClearResult("OdbcResult")

RAthena::dbClearResult("AthenaResult")

RJDBC::dbClearResult("JDBCResult")

RMariaDB::dbClearResult("MariaDBResult")

RMySQL::dbClearResult("MySQLResult")

RPostgres::dbClearResult("PqResult")

RPostgreSQL::dbClearResult("PostgreSQLResult")

RPresto::dbClearResult("PrestoResult")

RSQLite::dbClearResult("SQLiteResult")

sergeant::dbClearResult("DrillResult")

sparklyr::dbClearResult("DBISparkResult")

dbClearResult(res, ...)

Arguments

res

An object inheriting from DBIResult.

...

Other arguments passed on to methods.

Value

dbClearResult() returns TRUE, invisibly, for result sets obtained from dbSendQuery(), dbSendStatement(), or dbSendQueryArrow(),

The data retrieval flow

This section gives a complete overview over the flow for the execution of queries that return tabular data.

Most of this flow, except calling dbBind(), is implemented by dbGetQuery(), which should be sufficient unless you want to access the results in a paged way or you have a parameterized query. This flow requires an active connection established by dbConnect(). See also vignette("dbi-advanced") for a walkthrough.

Use dbSendQuery() to create a result set object of class DBIResult.

Optionally, bind query parameters with dbBind(). This is required only if the query contains placeholders such as ? or $1, depending on the database backend.

Optionally, use dbColumnInfo() to retrieve the structure of the result set without retrieving actual data.

Use dbFetch() to get the entire result set, a page of results, or the remaining rows. Fetching zero rows is also possible to retrieeve the structure of the result set as a data frame. This step can be called multiple times. Only forward paging is supported, you need to cache previous pages if you need to navigate backwards.

Use dbHasCompleted() to tell when you're done. This method returns TRUE if no more rows are available for fetching.

Use dbClearResult() to clean up the result set object. This step is mandatory even if no rows have been fetched or if an error has occurred during the processing. It is good practice to use on.exit() or withr::defer() to ensure that this step is always executed.

The command execution flow

This section gives a complete overview over the flow for the execution of SQL statements that have side effects such as stored procedures, inserting or deleting data, Most of this flow, except calling dbBind(), is implemented by dbExecute(), which should be sufficient for non-parameterized queries. This flow requires an active connection established by dbConnect(). See also vignette("dbi-advanced") for a walkthrough.

Use dbSendStatement() to create a result set object of class DBIResult. For some queries you need to pass immediate = TRUE.

Optionally, bind query parameters with dbBind(). This is required only if the query contains placeholders such as ? or $1, depending on the database backend.

Optionally, use dbGetRowsAffected() to retrieve the number of rows affected by the query.

Use dbClearResult() to clean up the result set object. This step is mandatory even if no rows have been fetched or if an error has occurred during the processing. It is good practice to use on.exit() or withr::defer() to ensure that this step is always executed.

Failure modes

An attempt to close an already closed result set issues a warning for dbSendQuery(), dbSendStatement(), and dbSendQueryArrow(),

Specification

dbClearResult() frees all resources associated with retrieving the result of a query or update operation. The DBI backend can expect a call to dbClearResult() for each dbSendQuery() or dbSendStatement() call.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

rs <- dbSendQuery(con, "SELECT 1")
print(dbFetch(rs))
#>   1
#> 1 1

dbClearResult(rs)
dbDisconnect(con)