dbConnect {DBI}	R Documentation

Create a connection to a DBMS

Description

Connect to a DBMS going through the appropriate authorization procedure. Some implementations may allow you to have multiple connections open, so you may invoke this function repeatedly assigning its output to different objects. The authorization mechanism is left unspecified, so check the documentation of individual drivers for details.

Usage

dbConnect(drv, ...)

Arguments

drv
an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

...
authorization arguments needed by the DBMS instance; these typically include user, password, dbname, host, port, etc. For details see the appropriate DBIDriver.

Details

Each driver will define what other arguments are required, e.g., "dbname" for the database name, "username", and "password".

Value

An object that extends DBIConnection in a database-specific manner. For instance dbConnect(RMySQL::MySQL()) produces an object of class MySQLConnection. This object is used to direct commands to the database engine.

See Also

dbDisconnect to disconnect from a database.

Other DBIDriver generics: DBIDriver-class, dbDataType, dbDriver, dbGetInfo, dbIsValid, dbListConnections

Examples

# SQLite only needs a path to the database. Other database drivers
# will require more details (like username, password, host, port etc)
con <- dbConnect(RSQLite::SQLite(), ":memory:")
con

dbListTables(con)

dbDisconnect(con)

[Package DBI version 0.5 Index]