Disconnect (close) a connection

This closes the connection, discards all pending work, and frees resources (e.g., memory, sockets).

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

AzureKusto::dbDisconnect("AzureKustoDriver")

bigrquery::dbDisconnect("BigQueryConnection")

DatabaseConnector::dbDisconnect("DatabaseConnectorConnection")

dittodb::dbDisconnect("DBIMockConnection")

duckdb::dbDisconnect("duckdb_connection")

implyr::dbDisconnect("src_impala")

lazysf::dbDisconnect("SFSQLConnection")

odbc::dbDisconnect("OdbcConnection")

pool::dbDisconnect("Pool")

RAthena::dbDisconnect("AthenaConnection")

RJDBC::dbDisconnect("JDBCConnection")

RMariaDB::dbDisconnect("MariaDBConnection")

RMySQL::dbDisconnect("MySQLConnection")

RPostgres::dbDisconnect("PqConnection")

RPostgreSQL::dbDisconnect("PostgreSQLConnection")

RPresto::dbDisconnect("PrestoConnection")

RSQLite::dbDisconnect("SQLiteConnection")

sergeant::dbDisconnect("DrillConnection")

sparklyr::dbDisconnect("spark_connection")

Usage

dbDisconnect(conn, ...)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

...

Other parameters passed on to methods.

Value

dbDisconnect() returns TRUE, invisibly.

Failure modes

A warning is issued on garbage collection when a connection has been released without calling dbDisconnect(), but this cannot be tested automatically. A warning is issued immediately when calling dbDisconnect() on an already disconnected or invalid connection.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbDisconnect(con)