Does a table exist?

Returns if a table given by name exists in the database.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

AzureKusto::dbExistsTable("AzureKustoConnection", "ANY")

bigrquery::dbExistsTable("BigQueryConnection", "character")

DatabaseConnector::dbExistsTable("DatabaseConnectorConnection", "character")

dittodb::dbExistsTable("DBIMockConnection", "character")

dittodb::dbExistsTable("DBIMockConnection", "Id")

duckdb::dbExistsTable("duckdb_connection", "character")

lazysf::dbExistsTable("SFSQLConnection", "ANY")

odbc::dbExistsTable("OdbcConnection", "character")

odbc::dbExistsTable("OdbcConnection", "Id")

odbc::dbExistsTable("OdbcConnection", "SQL")

pool::dbExistsTable("Pool", "ANY")

RAthena::dbExistsTable("AthenaConnection", "character")

RJDBC::dbExistsTable("JDBCConnection", "ANY")

RMariaDB::dbExistsTable("MariaDBConnection", "character")

RMySQL::dbExistsTable("MySQLConnection", "character")

RPostgres::dbExistsTable("PqConnection", "character")

RPostgres::dbExistsTable("PqConnection", "Id")

RPostgreSQL::dbExistsTable("PostgreSQLConnection", "character")

RPresto::dbExistsTable("PrestoConnection", "ANY")

RSQLite::dbExistsTable("SQLiteConnection", "character")

sparklyr::dbExistsTable("spark_connection", "character")

dbExistsTable(conn, name, ...)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

...

Other parameters passed on to methods.

Value

dbExistsTable() returns a logical scalar, TRUE if the table or view specified by the name argument exists, FALSE otherwise.

This includes temporary tables if supported by the database.

Failure modes

An error is raised when calling this method for a closed or invalid connection. An error is also raised if name cannot be processed with dbQuoteIdentifier() or if this results in a non-scalar.

Specification

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbExistsTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name)

If the result of a call to dbQuoteIdentifier(): no more quoting is done

For all tables listed by dbListTables(), dbExistsTable() returns TRUE.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbExistsTable(con, "iris")
#> [1] FALSE
dbWriteTable(con, "iris", iris)
dbExistsTable(con, "iris")
#> [1] TRUE

dbDisconnect(con)