Fetch records from a previously executed query

Fetch the next n elements (rows) from the result set and return them as a data.frame.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

RAthena::dbFetch("AthenaResult", "ANY")

AzureKusto::dbFetch("AzureKustoResult", "ANY")

bigrquery::dbFetch("BigQueryResult", "ANY")

RClickhouse::dbFetch("ClickhouseResult", "ANY")

DatabaseConnector::dbFetch("DatabaseConnectorResult", "ANY")

dittodb::dbFetch("DBIMockResult", "ANY")

sparklyr::dbFetch("DBISparkResult", "ANY")

sergeant::dbFetch("DrillResult", "ANY")

duckdb::dbFetch("duckdb_result", "ANY")

RMariaDB::dbFetch("MariaDBResult", "ANY")

RMySQL::dbFetch("MySQLResult", "missing")

RMySQL::dbFetch("MySQLResult", "numeric")

odbc::dbFetch("OdbcResult", "ANY")

RPostgres::dbFetch("PqResult", "ANY")

RPresto::dbFetch("PrestoResult", "integer")

RPresto::dbFetch("PrestoResult", "missing")

RPresto::dbFetch("PrestoResult", "numeric")

lazysf::dbFetch("SFSQLResult", "ANY")

RSQLite::dbFetch("SQLiteResult", "ANY")

dbFetch(res, n = -1, ...)

fetch(res, n = -1, ...)

Arguments

res

An object inheriting from DBIResult, created by dbSendQuery().

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

...

Other arguments passed on to methods.

Details

fetch() is provided for compatibility with older DBI clients - for all new code you are strongly encouraged to use dbFetch(). The default implementation for dbFetch() calls fetch() so that it is compatible with existing code. Modern backends should implement for dbFetch() only.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)

# Fetch all results
rs <- dbSendQuery(con, "SELECT * FROM mtcars WHERE cyl = 4")
dbFetch(rs)
#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 2  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 3  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 4  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 5  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 6  33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 7  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 8  27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 9  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 10 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 11 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
dbClearResult(rs)

# Fetch in chunks
rs <- dbSendQuery(con, "SELECT * FROM mtcars")
while (!dbHasCompleted(rs)) {
  chunk <- dbFetch(rs, 10)
  print(nrow(chunk))
}
#> [1] 10
#> [1] 10
#> [1] 10
#> [1] 2

dbClearResult(rs)
dbDisconnect(con)