Get connection arguments

Returns the arguments stored in a DBIConnector object for inspection, optionally evaluating them. This function is called by dbConnect() and usually does not need to be called directly.

dbGetConnectArgs(drv, eval = TRUE, ...)

Arguments

drv

A object inheriting from DBIConnector.

eval

Set to FALSE to return the functions that generate the argument instead of evaluating them.

...

Other arguments passed on to methods. Not otherwise used.

See also

Examples

cnr <- new("DBIConnector",
  .drv = RSQLite::SQLite(),
  .conn_args = list(dbname = ":memory:", password = function() "supersecret")
)
dbGetConnectArgs(cnr)
#> $dbname
#> [1] ":memory:"
#> 
#> $password
#> [1] "supersecret"
#> 
dbGetConnectArgs(cnr, eval = FALSE)
#> $dbname
#> [1] ":memory:"
#> 
#> $password
#> function() "supersecret"
#> <environment: 0x560d2ecdb720>
#>