dbGetInfo {DBI}	R Documentation

Get DBMS metadata

Description

Get DBMS metadata

Usage

dbGetInfo(dbObj, ...)

Arguments

dbObj
An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

...
Other arguments to methods.

Value

a named list

Implementation notes

For DBIDriver subclasses, this should include the version of the package (driver.version) and the version of the underlying client library (client.version).

For DBIConnection objects this should report the version of the DBMS engine (db.version), database name (dbname), username, (username), host (host), port (port), etc. It MAY also include any other arguments related to the connection (e.g., thread id, socket or TCP connection type). It MUST NOT include the password.

For DBIResult objects, this should include the statement being executed (statement), how many rows have been fetched so far (in the case of queries, row.count), how many rows were affected (deleted, inserted, changed, (rows.affected), and if the query is complete (has.completed).

The default implementation for DBIResult objects constructs such a list from the return values of the corresponding methods, dbGetStatement(), dbGetRowCount(), dbGetRowsAffected(), and dbHasCompleted().

See Also

Other DBIDriver generics: DBIDriver-class, dbCanConnect, dbConnect, dbDataType, dbDriver, dbIsReadOnly, dbIsValid, dbListConnections

Other DBIConnection generics: DBIConnection-class, dbAppendTable, dbCreateTable, dbDataType, dbDisconnect, dbExecute, dbExistsTable, dbGetException, dbGetQuery, dbIsReadOnly, dbIsValid, dbListFields, dbListObjects, dbListResults, dbListTables, dbReadTable, dbRemoveTable, dbSendQuery, dbSendStatement, dbWriteTable

Other DBIResult generics: DBIResult-class, dbBind, dbClearResult, dbColumnInfo, dbFetch, dbGetRowCount, dbGetRowsAffected, dbGetStatement, dbHasCompleted, dbIsReadOnly, dbIsValid, dbQuoteIdentifier, dbQuoteLiteral, dbQuoteString, dbUnquoteIdentifier

[Package DBI version 1.0.0 Index]