Get DBMS metadata

Retrieves information on objects of class DBIDriver, DBIConnection or DBIResult.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

adbi::dbGetInfo("AdbiConnection")

adbi::dbGetInfo("AdbiDriver")

bigrquery::dbGetInfo("BigQueryConnection")

bigrquery::dbGetInfo("BigQueryDriver")

DatabaseConnector::dbGetInfo("DatabaseConnectorConnection")

DatabaseConnector::dbGetInfo("DatabaseConnectorDriver")

dittodb::dbGetInfo("DBIMockConnection")

dittodb::dbGetInfo("DBIMockResult")

duckdb::dbGetInfo("duckdb_connection")

duckdb::dbGetInfo("duckdb_driver")

duckdb::dbGetInfo("duckdb_result")

lazysf::dbGetInfo("SFSQLDriver")

odbc::dbGetInfo("OdbcConnection")

odbc::dbGetInfo("OdbcDriver")

pool::dbGetInfo("Pool")

RAthena::dbGetInfo("AthenaConnection")

RAthena::dbGetInfo("AthenaResult")

RJDBC::dbGetInfo("JDBCConnection")

RJDBC::dbGetInfo("JDBCDriver")

RJDBC::dbGetInfo("JDBCResult")

RMariaDB::dbGetInfo("MariaDBConnection")

RMariaDB::dbGetInfo("MariaDBDriver")

RMySQL::dbGetInfo("MySQLConnection")

RMySQL::dbGetInfo("MySQLDriver")

RMySQL::dbGetInfo("MySQLResult")

RPostgres::dbGetInfo("PqConnection")

RPostgres::dbGetInfo("PqDriver")

RPostgreSQL::dbGetInfo("PostgreSQLConnection")

RPostgreSQL::dbGetInfo("PostgreSQLDriver")

RPostgreSQL::dbGetInfo("PostgreSQLResult")

RPresto::dbGetInfo("PrestoConnection")

RPresto::dbGetInfo("PrestoDriver")

RPresto::dbGetInfo("PrestoResult")

RSQLite::dbGetInfo("SQLiteConnection")

RSQLite::dbGetInfo("SQLiteDriver")

sergeant::dbGetInfo("DrillConnection")

sergeant::dbGetInfo("DrillDriver")

sparklyr::dbGetInfo("spark_connection")

Usage

dbGetInfo(dbObj, ...)

Arguments

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

...

Other arguments to methods.

Value

For objects of class DBIDriver, dbGetInfo()

returns a named list that contains at least the following components:

driver.version: the package version of the DBI backend,

client.version: the version of the DBMS client library.

For objects of class DBIConnection, dbGetInfo()

returns a named list that contains at least the following components:

db.version: version of the database server,

dbname: database name,

username: username to connect to the database,

host: hostname of the database server,

port: port on the database server. It must not contain a password component. Components that are not applicable should be set to NA.

For objects of class DBIResult, dbGetInfo()

returns a named list that contains at least the following components:

statatment: the statement used with dbSendQuery() or dbExecute(), as returned by dbGetStatement(),

row.count: the number of rows fetched so far (for queries), as returned by dbGetRowCount(),

rows.affected: the number of rows affected (for statements), as returned by dbGetRowsAffected()

has.completed: a logical that indicates if the query or statement has completed, as returned by dbHasCompleted().

Implementation notes

The default implementation for DBIResult objects constructs such a list from the return values of the corresponding methods, dbGetStatement(), dbGetRowCount(), dbGetRowsAffected(), and dbHasCompleted().

See also