Retrieve results from a query as an Arrow object

Returns the result of a query as an Arrow object. dbGetQueryArrow() comes with a default implementation (which should work with most backends) that calls dbSendQueryArrow(), then dbFetchArrow(), ensuring that the result is always freed by dbClearResult(). For passing query parameters, see dbSendQueryArrow(), in particular the "The data retrieval flow for Arrow streams" section. For retrieving results as a data frame, see dbGetQuery().

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

pool::dbGetQueryArrow("Pool")

Usage

dbGetQueryArrow(conn, statement, ...)

Arguments

conn

A DBI::DBIConnection object, as returned by dbConnect().

statement

a character string containing SQL.

...

Other parameters passed on to methods.

Value

dbGetQueryArrow() always returns an object coercible to a data.frame, with as many rows as records were fetched and as many columns as fields in the result set, even if the result is a single value or has one or zero rows.

Details

This method is for SELECT queries only (incl. other SQL statements that return a SELECT-alike result, e.g., execution of a stored procedure or data manipulation queries like INSERT INTO ... RETURNING ...). To execute a stored procedure that does not return a result set, use dbExecute().

Some backends may support data manipulation statements through this method. However, callers are strongly advised to use dbExecute() for data manipulation statements.

Implementation notes

Subclasses should override this method only if they provide some sort of performance optimization.

Failure modes

An error is raised when issuing a query over a closed or invalid connection, if the syntax of the query is invalid, or if the query is not a non-NA string. The object returned by dbGetQueryArrow() can also be passed to nanoarrow::as_nanoarrow_array_stream() to create a nanoarrow array stream object that can be used to read the result set in batches. The chunk size is implementation-specific.

Additional arguments

The following arguments are not part of the dbGetQueryArrow() generic (to improve compatibility across backends) but are part of the DBI specification:

params (default: NULL)

immediate (default: NULL)

They must be provided as named arguments. See the "Specification" and "Value" sections for details on their usage.

The param argument allows passing query parameters, see DBI::dbBind() for details.

Specification for the immediate argument

The immediate argument supports distinguishing between "direct" and "prepared" APIs offered by many database drivers. Passing immediate = TRUE leads to immediate execution of the query or statement, via the "direct" API (if supported by the driver). The default NULL means that the backend should choose whatever API makes the most sense for the database, and (if relevant) tries the other API if the first attempt fails. A successful second attempt should result in a message that suggests passing the correct immediate argument. Examples for possible behaviors:

DBI backend defaults to immediate = TRUE internally

A query without parameters is passed: query is executed

A query with parameters is passed:

params not given: rejected immediately by the database because of a syntax error in the query, the backend tries immediate = FALSE (and gives a message)

params given: query is executed using immediate = FALSE

DBI backend defaults to immediate = FALSE internally

A query without parameters is passed:

simple query: query is executed

"special" query (such as setting a config options): fails, the backend tries immediate = TRUE (and gives a message)

A query with parameters is passed:

params not given: waiting for parameters via DBI::dbBind()

params given: query is executed

See also

Examples

# Retrieve data as arrow table
con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)
dbGetQueryArrow(con, "SELECT * FROM mtcars")
#> <nanoarrow_array_stream struct<mpg: double, cyl: double, disp: double, hp: double, drat: double, wt: double, qsec: double, vs: double, am: double, gear: double, carb: double>>
#>  $ get_schema:function ()  
#>  $ get_next  :function (schema = x$get_schema(), validate = TRUE)  
#>  $ release   :function ()  

dbDisconnect(con)