dbGetRowsAffected {DBI}	R Documentation

The number of rows affected

Description

This function returns the number of rows that were added, deleted, or updated by data modifying query. For a selection query, this function returns 0.

Usage

dbGetRowsAffected(res, ...)

Arguments

res
An object inheriting from DBIResult.

...
Other arguments passed on to methods.

Value

a numeric vector of length 1

See Also

Other DBIResult generics: dbClearResult, dbColumnInfo, dbFetch, dbGetRowCount, dbGetStatement, dbHasCompleted

[Package DBI version 0.4-1 Index]