Get the statement associated with a result set

Returns the statement that was passed to dbSendQuery() or dbSendStatement().

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

adbi::dbGetStatement("AdbiResult")

adbi::dbGetStatement("AdbiResultArrow")

bigrquery::dbGetStatement("BigQueryResult")

DatabaseConnector::dbGetStatement("DatabaseConnectorDbiResult")

DatabaseConnector::dbGetStatement("DatabaseConnectorJdbcResult")

duckdb::dbGetStatement("duckdb_result")

odbc::dbGetStatement("OdbcResult")

RAthena::dbGetStatement("AthenaResult")

RMariaDB::dbGetStatement("MariaDBResult")

RMySQL::dbGetStatement("MySQLResult")

RPostgres::dbGetStatement("PqResult")

RPostgreSQL::dbGetStatement("PostgreSQLResult")

RPresto::dbGetStatement("PrestoResult")

RSQLite::dbGetStatement("SQLiteResult")

sergeant::dbGetStatement("DrillResult")

sparklyr::dbGetStatement("DBISparkResult")

Usage

dbGetStatement(res, ...)

Arguments

res

An object inheriting from DBI::DBIResult.

...

Other arguments passed on to methods.

Value

dbGetStatement() returns a string, the query used in either DBI::dbSendQuery() or DBI::dbSendStatement().

Failure modes

Attempting to query the statement for a result set cleared with DBI::dbClearResult() gives an error.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)
rs <- dbSendQuery(con, "SELECT * FROM mtcars")
dbGetStatement(rs)
#> [1] "SELECT * FROM mtcars"

dbClearResult(rs)
dbDisconnect(con)