Completion status

This method returns if the operation has completed. A SELECT query is completed if all rows have been fetched. A data manipulation statement is always completed.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

adbi::dbHasCompleted("AdbiResult")

adbi::dbHasCompleted("AdbiResultArrow")

bigrquery::dbHasCompleted("BigQueryResult")

DatabaseConnector::dbHasCompleted("DatabaseConnectorDbiResult")

DatabaseConnector::dbHasCompleted("DatabaseConnectorJdbcResult")

dittodb::dbHasCompleted("DBIMockResult")

duckdb::dbHasCompleted("duckdb_result")

lazysf::dbHasCompleted("SFSQLResult")

odbc::dbHasCompleted("OdbcResult")

RAthena::dbHasCompleted("AthenaResult")

RJDBC::dbHasCompleted("JDBCResult")

RMariaDB::dbHasCompleted("MariaDBResult")

RMySQL::dbHasCompleted("MySQLResult")

RPostgres::dbHasCompleted("PqResult")

RPostgreSQL::dbHasCompleted("PostgreSQLResult")

RPresto::dbHasCompleted("PrestoResult")

RSQLite::dbHasCompleted("SQLiteResult")

sergeant::dbHasCompleted("DrillResult")

sparklyr::dbHasCompleted("DBISparkResult")

Usage

dbHasCompleted(res, ...)

Arguments

res

An object inheriting from DBI::DBIResult.

...

Other arguments passed on to methods.

Value

dbHasCompleted() returns a logical scalar. For a query initiated by DBI::dbSendQuery() with non-empty result set, dbHasCompleted() returns FALSE initially and TRUE after calling DBI::dbFetch() without limit. For a query initiated by DBI::dbSendStatement(), dbHasCompleted() always returns TRUE.

The data retrieval flow

This section gives a complete overview over the flow for the execution of queries that return tabular data as data frames.

Most of this flow, except repeated calling of dbBind() or dbBindArrow(), is implemented by dbGetQuery(), which should be sufficient unless you want to access the results in a paged way or you have a parameterized query that you want to reuse. This flow requires an active connection established by dbConnect(). See also vignette("dbi-advanced") for a walkthrough.

Use dbSendQuery() to create a result set object of class DBIResult.

Optionally, bind query parameters with dbBind() or dbBindArrow(). This is required only if the query contains placeholders such as ? or $1, depending on the database backend.

Optionally, use dbColumnInfo() to retrieve the structure of the result set without retrieving actual data.

Use dbFetch() to get the entire result set, a page of results, or the remaining rows. Fetching zero rows is also possible to retrieve the structure of the result set as a data frame. This step can be called multiple times. Only forward paging is supported, you need to cache previous pages if you need to navigate backwards.

Use dbHasCompleted() to tell when you're done. This method returns TRUE if no more rows are available for fetching.

Repeat the last four steps as necessary.

Use dbClearResult() to clean up the result set object. This step is mandatory even if no rows have been fetched or if an error has occurred during the processing. It is good practice to use on.exit() or withr::defer() to ensure that this step is always executed.

Failure modes

Attempting to query completion status for a result set cleared with DBI::dbClearResult() gives an error.

Specification

The completion status for a query is only guaranteed to be set to FALSE after attempting to fetch past the end of the entire result. Therefore, for a query with an empty result set, the initial return value is unspecified, but the result value is TRUE after trying to fetch only one row.

Similarly, for a query with a result set of length n, the return value is unspecified after fetching n rows, but the result value is TRUE after trying to fetch only one more row.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)
rs <- dbSendQuery(con, "SELECT * FROM mtcars")

dbHasCompleted(rs)
#> [1] FALSE
ret1 <- dbFetch(rs, 10)
dbHasCompleted(rs)
#> [1] FALSE
ret2 <- dbFetch(rs)
dbHasCompleted(rs)
#> [1] TRUE

dbClearResult(rs)
dbDisconnect(con)