Is this DBMS object read only?

This generic tests whether a database object is read only.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

pool::dbIsReadOnly("Pool")

Usage

dbIsReadOnly(dbObj, ...)

Arguments

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

...

Other arguments to methods.

See also

Examples

dbIsReadOnly(ANSI())
#> [1] FALSE