dbListConnections {DBI}	R Documentation

List currently open connections

Description

DEPRECATED, drivers are no longer required to implement this method. Keep track of the connections you opened if you require a list.

Usage

dbListConnections(drv, ...)

Arguments

drv
A object inheriting from DBIDriver

...
Other arguments passed on to methods.

Value

a list

See Also

Other DBIDriver generics: DBIDriver-class, dbCanConnect(), dbConnect(), dbDataType(), dbDriver(), dbGetInfo(), dbIsReadOnly(), dbIsValid()

[Package DBI version 1.1.3 Index]