List currently open connections

DEPRECATED, drivers are no longer required to implement this method. Keep track of the connections you opened if you require a list.

dbListConnections(drv, ...)

Arguments

drv
A object inheriting from DBIDriver

...
Other arguments passed on to methods.

Value

a list

See also