List field names of a remote table

Returns the field names of a remote table as a character vector.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

adbi::dbListFields("AdbiConnection", "character")

adbi::dbListFields("AdbiConnection", "Id")

adbi::dbListFields("AdbiConnection", "SQL")

AzureKusto::dbListFields("AzureKustoConnection", "character")

bigrquery::dbListFields("BigQueryConnection", "AsIs")

bigrquery::dbListFields("BigQueryConnection", "character")

bigrquery::dbListFields("BigQueryConnection", "Id")

DatabaseConnector::dbListFields("DatabaseConnectorConnection", "character")

dittodb::dbListFields("DBIMockConnection", "ANY")

dittodb::dbListFields("DBIMockConnection", "character")

dittodb::dbListFields("DBIMockConnection", "Id")

duckdb::dbListFields("duckdb_connection", "character")

odbc::dbListFields("OdbcConnection", "character")

odbc::dbListFields("OdbcConnection", "Id")

odbc::dbListFields("OdbcConnection", "SQL")

pool::dbListFields("Pool", "ANY")

RAthena::dbListFields("AthenaConnection", "character")

RJDBC::dbListFields("JDBCConnection", "ANY")

RMySQL::dbListFields("MySQLConnection", "character")

RMySQL::dbListFields("MySQLResult", "missing")

RPostgres::dbListFields("PqConnection", "character")

RPostgres::dbListFields("PqConnection", "Id")

RPostgreSQL::dbListFields("PostgreSQLConnection", "character")

RPostgreSQL::dbListFields("PostgreSQLResult", "missing")

RPresto::dbListFields("PrestoConnection", "character")

RPresto::dbListFields("PrestoConnection", "dbplyr_schema")

RPresto::dbListFields("PrestoConnection", "Id")

RPresto::dbListFields("PrestoConnection", "SQL")

RPresto::dbListFields("PrestoResult", "missing")

sergeant::dbListFields("DrillConnection", "character")

sergeant::dbListFields("DrillResult", "missing")

Usage

dbListFields(conn, name, ...)

Arguments

conn

A DBI::DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

...

Other parameters passed on to methods.

Value

dbListFields() returns a character vector that enumerates all fields in the table in the correct order. This also works for temporary tables if supported by the database. The returned names are suitable for quoting with dbQuoteIdentifier().

Failure modes

If the table does not exist, an error is raised. Invalid types for the name argument (e.g., character of length not equal to one, or numeric) lead to an error. An error is also raised when calling this method for a closed or invalid connection.

Specification

The name argument can be

a string

the return value of DBI::dbQuoteIdentifier()

a value from the table column from the return value of DBI::dbListObjects() where is_prefix is FALSE

A column named row_names is treated like any other column.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)
dbListFields(con, "mtcars")
#>  [1] "mpg"  "cyl"  "disp" "hp"   "drat" "wt"   "qsec" "vs"   "am"   "gear"
#> [11] "carb"

dbDisconnect(con)