List remote tables

Returns the unquoted names of remote tables accessible through this connection. This should include views and temporary objects, but not all database backends (in particular RMariaDB and RMySQL) support this.

Methods in other packages

RSQLite::dbListTables("SQLiteConnection")

dbListTables(conn, ...)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

...
Other parameters passed on to methods.

Value

dbListTables() returns a character vector that enumerates all tables and views in the database. Tables added with dbWriteTable() are part of the list, including temporary tables if supported by the database. As soon a table is removed from the database, it is also removed from the list of database tables.

The returned names are suitable for quoting with dbQuoteIdentifier(). An error is raised when calling this method for a closed or invalid connection.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbListTables(con)

#> character(0)

dbWriteTable(con, "mtcars", mtcars)
dbListTables(con)

#> [1] "mtcars"

dbDisconnect(con)