List remote tables

Returns the unquoted names of remote tables accessible through this connection. This should include views and temporary objects, but not all database backends (in particular RMariaDB and RMySQL) support this.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

adbi::dbListTables("AdbiConnection")

AzureKusto::dbListTables("AzureKustoConnection")

bigrquery::dbListTables("BigQueryConnection")

DatabaseConnector::dbListTables("DatabaseConnectorConnection")

dittodb::dbListTables("DBIMockConnection")

duckdb::dbListTables("duckdb_connection")

lazysf::dbListTables("SFSQLConnection")

odbc::dbListTables("Microsoft SQL Server")

odbc::dbListTables("OdbcConnection")

pool::dbListTables("Pool")

RAthena::dbListTables("AthenaConnection")

RJDBC::dbListTables("JDBCConnection")

RMariaDB::dbListTables("MariaDBConnection")

RMySQL::dbListTables("MySQLConnection")

RPostgres::dbListTables("PqConnection")

RPostgreSQL::dbListTables("PostgreSQLConnection")

RPresto::dbListTables("PrestoConnection")

RSQLite::dbListTables("SQLiteConnection")

sparklyr::dbListTables("spark_connection")

Usage

dbListTables(conn, ...)

Arguments

conn

A DBI::DBIConnection object, as returned by dbConnect().

...

Other parameters passed on to methods.

Value

dbListTables() returns a character vector that enumerates all tables and views in the database. Tables added with DBI::dbWriteTable() are part of the list. As soon a table is removed from the database, it is also removed from the list of database tables.

The same applies to temporary tables if supported by the database.

The returned names are suitable for quoting with dbQuoteIdentifier().

Failure modes

An error is raised when calling this method for a closed or invalid connection.

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbListTables(con)
#> character(0)
dbWriteTable(con, "mtcars", mtcars)
dbListTables(con)
#> [1] "mtcars"

dbDisconnect(con)