Read database tables as data frames

Reads a database table to a data frame, optionally converting a column to row names and converting the column names to valid R identifiers. Use dbReadTableArrow() instead to obtain an Arrow object.

Methods in other packages

This documentation page describes the generics. Refer to the documentation pages linked below for the documentation for the methods that are implemented in various backend packages.

AzureKusto::dbReadTable("AzureKustoConnection", "character")

bigrquery::dbReadTable("BigQueryConnection", "AsIs")

bigrquery::dbReadTable("BigQueryConnection", "character")

bigrquery::dbReadTable("BigQueryConnection", "Id")

DatabaseConnector::dbReadTable("DatabaseConnectorConnection", "character")

lazysf::dbReadTable("SFSQLConnection", "character")

pool::dbReadTable("Pool", "ANY")

RJDBC::dbReadTable("JDBCConnection", "ANY")

RJDBC::dbReadTable("JDBCConnection", "character")

RMariaDB::dbReadTable("MariaDBConnection", "character")

RMySQL::dbReadTable("MySQLConnection", "character")

RPostgres::dbReadTable("PqConnection", "character")

RPostgreSQL::dbReadTable("PostgreSQLConnection", "character")

RPresto::dbReadTable("PrestoConnection", "ANY")

RPresto::dbReadTable("PrestoConnection", "character")

RPresto::dbReadTable("PrestoConnection", "dbplyr_schema")

RPresto::dbReadTable("PrestoConnection", "Id")

RPresto::dbReadTable("PrestoConnection", "SQL")

RSQLite::dbReadTable("SQLiteConnection", "character")

sparklyr::dbReadTable("spark_connection", "character")

Usage

dbReadTable(conn, name, ...)

Arguments

conn

A DBI::DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

...

Other parameters passed on to methods.

Value

dbReadTable() returns a data frame that contains the complete data from the remote table, effectively the result of calling DBI::dbGetQuery() with SELECT * FROM <name>.

An empty table is returned as a data frame with zero rows.

The presence of rownames depends on the row.names argument, see DBI::sqlColumnToRownames() for details:

If FALSE or NULL, the returned data frame doesn't have row names.

If TRUE, a column named "row_names" is converted to row names.

If NA, a column named "row_names" is converted to row names if it exists, otherwise no translation occurs.

If a string, this specifies the name of the column in the remote table that contains the row names.

The default is row.names = FALSE.

If the database supports identifiers with special characters, the columns in the returned data frame are converted to valid R identifiers if the check.names argument is TRUE, If check.names = FALSE, the returned table has non-syntactic column names without quotes.

Details

This function returns a data frame. Use dbReadTableArrow() to obtain an Arrow object.

Failure modes

An error is raised if the table does not exist.

An error is raised if row.names is TRUE and no "row_names" column exists,

An error is raised if row.names is set to a string and no corresponding column exists.

An error is raised when calling this method for a closed or invalid connection. An error is raised if name cannot be processed with DBI::dbQuoteIdentifier() or if this results in a non-scalar. Unsupported values for row.names and check.names (non-scalars, unsupported data types, NA for check.names) also raise an error.

Additional arguments

The following arguments are not part of the dbReadTable() generic (to improve compatibility across backends) but are part of the DBI specification:

row.names (default: FALSE)

check.names

They must be provided as named arguments. See the "Value" section for details on their usage.

Specification

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbReadTable() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name)

If the result of a call to DBI::dbQuoteIdentifier(): no more quoting is done

See also

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars[1:10, ])
dbReadTable(con, "mtcars")
#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4

dbDisconnect(con)