sqlAppendTable {DBI}	R Documentation

Insert rows into a table

Description

sqlAppendTable generates a single SQL string that inserts a data frame into an existing table. sqlAppendTableTemplate generates a template suitable for use with dbBind(). These methods are mostly useful for backend implementers.

Usage

sqlAppendTable(con, table, values, row.names = NA, ...)

sqlAppendTableTemplate(con, table, values, row.names = NA, prefix = "?",
  ...)

Arguments

con
A database connection.

table
Name of the table. Escaped with dbQuoteIdentifier().

values
A data frame. Factors will be converted to character vectors. Character vectors will be escaped with dbQuoteString().

row.names
Either TRUE, FALSE, NA or a string.

If TRUE, always translate row names to a column called "row_names". If FALSE, never translate row names. If NA, translate rownames only if they're a character vector.

A string is equivalent to TRUE, but allows you to override the default name.

For backward compatibility, NULL is equivalent to FALSE.

...
Other arguments used by individual methods.

prefix
Parameter prefix to put in front of column id.

Examples

sqlAppendTable(ANSI(), "iris", head(iris))

sqlAppendTable(ANSI(), "mtcars", head(mtcars))
sqlAppendTable(ANSI(), "mtcars", head(mtcars), row.names = FALSE)
sqlAppendTableTemplate(ANSI(), "iris", iris)

sqlAppendTableTemplate(ANSI(), "mtcars", mtcars)
sqlAppendTableTemplate(ANSI(), "mtcars", mtcars, row.names = FALSE)

[Package DBI version 0.8 Index]