sqlData {DBI}	R Documentation

Convert a data frame into form suitable for upload to a SQL database

Description

This is a generic method that coerces R objects into vectors suitable for upload to the database. The output will vary a little from method to method depending on whether the main upload device is through a single SQL string or multiple parameterised queries. This method is mostly useful for backend implementers.

Usage

sqlData(con, value, row.names = NA, ...)

Arguments

con
A database connection.

value
A data frame

row.names
Either TRUE, FALSE, NA or a string.

If TRUE, always translate row names to a column called "row_names". If FALSE, never translate row names. If NA, translate rownames only if they're a character vector.

A string is equivalent to TRUE, but allows you to override the default name.

For backward compatibility, NULL is equivalent to FALSE.

...
Other arguments used by individual methods.

Details

The default method:

Converts factors to characters

Quotes all strings

Converts all columns to strings

Replaces NA with NULL

Examples

con <- dbConnect(RSQLite::SQLite(), ":memory:")

sqlData(con, head(iris))
sqlData(con, head(mtcars))

dbDisconnect(con)

[Package DBI version 0.5 Index]